+++ SwiftLabEffective/Models/Category.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Category {
+    var imageName: String
+    var name: String
+}
