+++ SwiftLabEffective/Models/Doctor.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+struct Doctor {
+    var id: UUID = UUID()
+    var imageName: String
+    var name: String
+    var specialization: String
+    var distance: Float
+    var mark: Float
+    var reviewsCount: Int
+    var openAtTime: String
+}
+
