+++ SwiftLabEffective/Models/ModelData.swift
@@ -1,0 +1,15 @@
+//
+//  ModelData.swift
+//  SwiftLabEffective
+//
+//  Created by  user on 13.11.2023.
+//
+
+import Foundation
+
+
+struct ModelData{
+    
+    
+    
+}
