+++ SwiftLabEffective/Models/User.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct User {
+    var id: UUID = UUID()
+    var imageAvatarName: String
+    var name: String
+}
