+++ SwiftLabEffective/Views/CategoryView.swift
@@ -1,38 +1,4 @@
-//
-//  CategoryView.swift
-//  SwiftLabEffective
-//
-//  Created by  user on 13.11.2023.
-//
-
 import SwiftUI
-
-struct Category {
-    var imageName: String
-    var name: String
-}
-
-struct CategoryRowView: View {
-    
-    let categories: [Category] = [
-        Category(imageName: "iconCovid", name: "Covid 19"),
-        Category(imageName: "iconDoctor", name: "Doctor"),
-        Category(imageName: "iconMedicine", name: "Medicine"),
-        Category(imageName: "iconHospital", name: "Hospital")
-    ]
-    
-    var body: some View {
-        HStack  {
-            ForEach(categories.indices) { index in
-                CategoryView(category: categories[index])
-                
-                if index != categories.indices.last{
-                    Spacer()
-                }
-            }
-        }
-    }
-}
 
 struct CategoryView: View {
     var category: Category
@@ -61,5 +27,8 @@ struct CategoryView: View {
 }
 
 #Preview {
-    CategoryRowView()
+    let category: Category =
+        Category(imageName: "iconCovid", name: "Covid 19")
+    
+    CategoryView(category: category)
 }
