+++ SwiftLabEffective/Views/WelcomeView.swift
@@ -1,23 +1,19 @@
-//
-//  WelcomeView.swift
-//  SwiftLabEffective
-//
-//  Created by  user on 13.11.2023.
-//
-
 import SwiftUI
 
 struct WelcomeView: View {
+    
+    let user: User = User(imageAvatarName: "avatar", name: "James")
+    
     var body: some View {
         HStack{
             VStack (alignment: .leading) {
                 Text("Hello,")
                     .font(Font.custom(FontTheme.poppinsRegular, size: 16))
-                    .foregroundColor(Color("myGrayColor"))
-                Text("Hi James").font(Font.custom(FontTheme.poppinsBold, size: 20))
+                    .foregroundColor(ColorTheme.Text.Secondary)
+                Text("Hi " + user.name).font(Font.custom(FontTheme.poppinsBold, size: 20))
             }
             Spacer()
-            Image("avatar")
+            Image(user.imageAvatarName)
                 .resizable()
                 .frame(width: 56, height: 56)
         }
